first pass on rect functionality

+++ resource/ui_madness.h
@@ -477,9 +477,8 @@ MAPI void madness_image_handle(Texture_Handle handle);
 MAPI void madness_ui_slider_scroll(String id, float* slider_val, float min, float max);
 MAPI void madness_ui_slider_arrow(String id, float* slider_val, float min, float max);
 
-//TODO: these should really check for overflows, there are toolchain specific overflow builtin functions
 MAPI bool madness_ui_scalar(String text, UI_Scalar_Type type, void* data, f64 value_change);
-MAPI bool madness_ui_scalar_range(String text, UI_Scalar_Type type, void* data, u64 increment_value, u64 min, u64 max);
+MAPI bool madness_ui_scalar_range(String text, UI_Scalar_Type type, void* data, u64 increment_value, f64 min, f64 max);
 
 MAPI bool madness_ui_u8(String text, u8* i, u32 increment_value);
 MAPI bool madness_ui_u16(String text, u16* i, u32 increment_value);
